shop

+++ ex/src/main/java/com/example/ex/model/repository/GenreRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
 @Repository
-public interface GenreRepository extends JpaRepository<Genre, Integer> {
+public interface GenreRepository extends JpaRepository<Genre, Long> {
 
     List<Genre> findByNameContainingIgnoreCaseOrderByName(String name);
+
 }
